pyproject.toml
[project]
name = "zettlr-rag"
version = "0.3.0"
description = "RAG for my academic papers library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index-llms-google-genai",
    "llama-index-embeddings-google-genai",
    "llama-index-vector-stores-chroma",
    "chromadb",
    "python-frontmatter",
    "watchdog",
    "python-dotenv",
    "pandas",
    "nest-asyncio",
]
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "black",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (undefined names, unused imports, etc.)
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear (common bug patterns)
    "C4",  # flake8-comprehensions (cleaner comprehensions)
    "UP",  # pyupgrade (modernize Python syntax)
    "N",   # pep8-naming (naming conventions)
    "RUF", # Ruff-specific rules
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_generics = true
warn_return_any = true
